Minor type fixes on value types

+++ MindTheWorldServer.Database/Tables/OilConsumption.sql
@@ -2,7 +2,7 @@
 (
 	[CountryId] INT NOT NULL , 
     [Year] INT NOT NULL, 
-    [Value] BIGINT NULL, 
+    [Value] DECIMAL(18, 6) NULL, 
 
     CONSTRAINT PK_OC_ID PRIMARY KEY ([CountryId], [Year]),
     CONSTRAINT FK_OC_CountryId FOREIGN KEY (CountryId) REFERENCES Country(Id)
